requirements.txt
Pillow~=10.2.0
reportlab~=4.0.9
